% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_static.R
\name{run_static}
\alias{run_static}
\title{run_static}
\usage{
run_static(sim_data, nC = 3, nI = 30000, nB = 10000, nT = 20)
}
\arguments{
\item{sim_data}{Simulated data from sim_data_static()}

\item{nC}{Number of chains}

\item{nI}{Number of iterations}

\item{nB}{Number of iterations to discard burn-in}

\item{nT}{Thinning rate}
}
\description{
Run static correlated-detection N-mixture model
}

