% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_dyn.R
\name{sim_data_dyn}
\alias{sim_data_dyn}
\title{sim_data_dyn}
\usage{
sim_data_dyn(nRoutes = 150, nStops = 50, nYears = 10, alpha0 = -0.5,
  alpha1 = -0.3, alpha2 = 0.2, beta0 = 2, sigma1 = 1, gamma = 0.35,
  omega = 0.6, sigma2 = 0.1, theta = c(0.3, 0.75), nSum = NULL)
}
\arguments{
\item{nRoutes}{Number of sample units}

\item{nStops}{Number of stops within each sample unit}

\item{nYears}{Number of years}

\item{alpha0}{Intercept for detection model}

\item{alpha1}{Linear effect of stop number on detection probability}

\item{alpha2}{Quadratic effect of stop number on detection probability}

\item{beta0}{Log expected number of individuals at each sampling unit in year 1}

\item{sigma1}{Standard deviation around mean count in year 1}

\item{gamma}{Recruitment rate}

\item{omega}{Survival rate}

\item{sigma2}{Standard deviation of random immigration from other sies}

\item{theta}{Vector containing the correlation terms}

\item{nSum}{Optional number of stops to aggregate}
}
\description{
Simulation of multi-season count data for the dynamic correlated-detection N-mixture model
}

