% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_static.R
\name{sim_data_static}
\alias{sim_data_static}
\title{sim_data_static}
\usage{
sim_data_static(nRoutes = 150, nStops = 50, alpha0 = -0.5,
  alpha1 = -0.3, alpha2 = 0.2, beta0 = 2, sigma = 1, theta = c(0.3,
  0.75), nSum = NULL)
}
\arguments{
\item{nRoutes}{Number of sample units}

\item{nStops}{Number of stops within each sample unit}

\item{alpha0}{Intercept for detection model}

\item{alpha1}{Linear effect of stop number on detection probability}

\item{alpha2}{Quadratic effect of stop number on detection probability}

\item{beta0}{Log expected number of individuals at each sampling unit}

\item{sigma}{Standard deviation around mean count}

\item{theta}{Vector containing the correlation terms}

\item{nSum}{Optional number of stops to aggregate}
}
\description{
Simulation of single-season count data for the static correlated-detection N-mixture model
}

